Separated testcode and mastercode into individual directories.
Code has not been confirmed working yet.
