+++ TeamCode_11353/src/main/java/org/firstinspires/ftc/teamcode/mastercode/UltimategoalHardware.java
@@ -28,7 +28,7 @@
  */
 
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.mastercode;
 
 import com.qualcomm.hardware.bosch.BNO055IMU;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
