+++ TeamCode_11353/src/main/java/org/firstinspires/ftc/teamcode/testcode/UltimateGoalTestHardwareMap.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.test_teamcode;
+package org.firstinspires.ftc.teamcode.testcode;
 
 import com.qualcomm.hardware.bosch.BNO055IMU;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
